AR_Soft/requirements.txt
asgiref==3.8.1
Django==4.2.21
sqlparse==0.5.1
celery==5.3.6
billiard==4.2.0
kombu==5.3.5
redis==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.1
django-celery-beat==2.6.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.34
django-cors-headers==4.7.0
